Simplify Figma JSON component parsing
Simplified process of parsing Figma json for new components.  Add to TYPES whenever we create a new component.  Each TYPES component that is found within the Figma JSON will be added to the final list of children.

This replaces the messy, exponentially increasing list of if statements.
No longer have to fiddle much with the code to add new components.